requirements.txt
absl-py==1.2.0
asttokens==2.0.5
astunparse==1.6.3
atomicwrites==1.4.1
attrs==22.1.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.11.1
black==22.6.0
brotlipy==0.7.0
cachetools==5.2.0
certifi 
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
cycler==0.11.0
debugpy==1.6.2
decorator==5.1.1
dnspython==2.2.1
email-validator==1.2.1
entrypoints==0.4
executing==0.9.1
filelock==3.7.1
Flask==2.1.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
flatbuffers==1.12
fonttools==4.34.4
gast==0.4.0
glob2==0.7
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.48.0
h5py==3.7.0
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
ipykernel==6.15.1
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jupyter-client==7.3.4
jupyter-core==4.11.1
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libarchive-c==4.0
libclang==14.0.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nest-asyncio==1.5.5
nltk==3.7
numexpr==2.8.3
numpy==1.23.1
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.2.0
pkginfo==1.8.3
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycosat==0.6.3
pycparser==2.21
Pygments==2.12.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pywin32==304
PyYAML==6.0
pyzmq==23.2.0
regex==2022.7.25
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.1
scipy==1.9.0
seaborn==0.11.2
six==1.16.0
sklearn==0.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
stack-data==0.3.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
tweepy==4.10.0
typed-ast==1.5.4
typing_extensions==4.3.0
urllib3==1.26.11
wcwidth==0.2.5
Werkzeug==2.2.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.8.1
